add fast_restart command
to node: when restarting the map, spawned bots will stay but won't autojoin a team.
added bot_team which will check every 3 sec if bot is spectator then select team if is.

also when fast_restarting at the end of the match when scoreboard is visible the scoreboard will persist and won't go away.